+++ test/array/index.js
@@ -2,9 +2,11 @@
 
 module.exports = {
     every: require('./every'),
+    fill: require('./fill'),
     filter: require('./filter'),
     forEach: require('./forEach'),
     map: require('./map'),
     reduce: require('./reduce'),
+    reverse: require('./reverse'),
     some: require('./some')
 };